requirements.txt
azure-cognitiveservices-speech
flask
openai
uuid
requests
opencv-python-headless
numpy
deepface
pillow
